# Improve encoding performance