From
```
BM_Assembler_EmitSingle_0_Operands      0.530 us        0.544 us      1120000 Instructions=1.83795M/s
BM_Assembler_EmitSingle_1_Operands      0.602 us        0.558 us      1120000 Instructions=1.792M/s
BM_Assembler_EmitSingle_2_Operands      0.687 us        0.670 us      1120000 Instructions=1.49333M/s
BM_Assembler_EmitSingle_3_Operands      0.827 us        0.837 us       896000 Instructions=1.19467M/s
BM_Assembler_EmitAll                     8145 us         8160 us           90 Instructions=1.23791M/s
BM_Serialization/4096                    1.85 ms         2.01 ms          264 BytesEncoded=13.4963M/s Instructions=2.03547M/s
BM_Serialization/8192                    4.25 ms         4.19 ms          179 BytesEncoded=13.864M/s Instructions=1.95516M/s
BM_Serialization/16384                   8.59 ms         8.96 ms           75 BytesEncoded=12.9678M/s Instructions=1.82891M/s
BM_Serialization/32768                   17.2 ms         18.4 ms           45 BytesEncoded=12.5989M/s Instructions=1.7806M/s
BM_Serialization/65536                   34.5 ms         34.2 ms           21 BytesEncoded=13.6657M/s Instructions=1.91479M/s
BM_Serialization/131072                  70.0 ms         67.7 ms            9 BytesEncoded=13.8664M/s Instructions=1.93583M/s
BM_Serialization/262144                   139 ms          138 ms            5 BytesEncoded=13.6542M/s Instructions=1.9065M/s
BM_Serialization/524288                   280 ms          281 ms            2 BytesEncoded=13.3502M/s Instructions=1.86414M/s
BM_Serialization/1048576                  567 ms          578 ms            1 BytesEncoded=12.9892M/s Instructions=1.81375M/s
BM_Serialization/2097152                 1127 ms         1125 ms            1 BytesEncoded=13.35M/s Instructions=1.86414M/s
```
to
```
BM_Assembler_EmitSingle_0_Operands      0.512 us        0.502 us      1120000 Instructions=1.99111M/s
BM_Assembler_EmitSingle_1_Operands      0.594 us        0.645 us       896000 Instructions=1.54984M/s
BM_Assembler_EmitSingle_2_Operands      0.695 us        0.698 us       896000 Instructions=1.4336M/s
BM_Assembler_EmitSingle_3_Operands      0.838 us        0.753 us       746667 Instructions=1.32741M/s
BM_Assembler_EmitAll                     8295 us         8507 us           90 Instructions=1.18738M/s
BM_Serialization/4096                    1.25 ms         1.23 ms          747 BytesEncoded=22.0069M/s Instructions=3.31901M/s
BM_Serialization/8192                    2.72 ms         2.70 ms          249 BytesEncoded=21.5281M/s Instructions=3.03599M/s
BM_Serialization/16384                   5.53 ms         5.47 ms          100 BytesEncoded=21.2425M/s Instructions=2.99593M/s
BM_Serialization/32768                   11.1 ms         11.7 ms           56 BytesEncoded=19.785M/s Instructions=2.7962M/s
BM_Serialization/65536                   22.6 ms         25.6 ms           25 BytesEncoded=18.2527M/s Instructions=2.5575M/s
BM_Serialization/131072                  45.6 ms         45.8 ms           15 BytesEncoded=20.4844M/s Instructions=2.85975M/s
BM_Serialization/262144                  91.9 ms         93.8 ms            7 BytesEncoded=20.0261M/s Instructions=2.7962M/s
BM_Serialization/524288                   183 ms          184 ms            4 BytesEncoded=20.4514M/s Instructions=2.8557M/s
BM_Serialization/1048576                  366 ms          367 ms            2 BytesEncoded=20.4511M/s Instructions=2.8557M/s
BM_Serialization/2097152                  736 ms          734 ms            1 BytesEncoded=20.4511M/s Instructions=2.8557M/s
```